废除旧的对router和gorm的处理方法，替换为新的自动化方法
