+++ server/utils/ast/ast_auto_enter_test.go
@@ -1,0 +1,7 @@
+package ast
+
+import "testing"
+
+func TestImportForAutoEnter(t *testing.T) {
+	ImportForAutoEnter("D:\\gin-vue-admin\\server\\api\\v1\\test\\enter.go", "ApiGroup", "test")
+}
